Always use base_class for STI models
When an instance of a model is an STI class, we can always load the instance using the base class's `find`er method. However, if you use a child class from the base STI class, then the `type` will be filtered, possibly preventing the model from being loaded as expected.